+++ app/policies/level_b_policy.rb
@@ -1,0 +1,9 @@
+class LevelBPolicy < ApplicationPolicy
+  def activity_upload?
+    return true if beis_user?
+  end
+
+  def budget_upload?
+    return true if beis_user?
+  end
+end